refactor controller rsync
